problems with recursive binOP
